# docs/developers/crm-server-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "App Connect Connector Server API",
    "version": "1.7.40",
    "description": "HTTP API exposed by an App Connect connector server (called the CRM or adapter server in older code). The configured manifest serverUrl is the HTTPS base URL, without a trailing slash, to which App Connect appends these paths. The contract combines shared @app-connect/core routes with connector-specific routes implemented by this server; fields and supported capabilities can vary by connector where an operation delegates to a connector interface.\n\nAuthorization: userJwt is an App Connect CRM-session JWT issued by /apiKeyLogin or /oauth-callback; it is not a RingCentral OAuth token. X-RC-Access-Token carries a RingCentral OAuth credential for RingCentral user or administrator validation. Legacy jwtToken and rcAccessToken query parameters remain active for older clients, but credentials in URLs can leak through logs and browser history and should be avoided.\n\nLegacy response convention: some connector operations return HTTP 200 with successful: false and returnMessage when the request was processed but the CRM operation failed. Callers must inspect the response body rather than treating every 2xx response as business success. Payloads can contain CRM records, communications content, recording URLs, and access tokens; do not log credentials or retain personal data beyond the workflow that requires it.",
    "contact": {
...
    },
    "license": {
...
  },
...
  "servers": [
...
  "security": [],
  "paths": {
...
    "/googleSheets/filePicker": {
      "get": {
        "tags": [
          "Google Integrations"
        ],
        "summary": "Open Google Sheets file picker",
        "operationId": "getGoogleSheetsFilePicker",
        "responses": {
          "200": {
            "description": "Google Sheets picker HTML page.",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "filePickerJwt": []
          }
        ],
        "parameters": [],
        "description": "Serves the bundled Google Sheets picker HTML for the App Connect user identified by the legacy token query parameter. This connector-specific route is omitted from the public API specification."
      }
    },
    "/googleSheets/sheet": {
      "post": {
        "tags": [
          "Google Integrations"
        ],
        "summary": "Create a Google Sheet",
        "operationId": "postGoogleSheetsSheet",
        "requestBody": {
          "$ref": "#/components/requestBodies/GenericJson"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/OkJson"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "legacyUserJwt": []
          }
        ],
        "description": "Creates a new Google Sheet for the App Connect user and returns its name and URL. Selecting an existing sheet uses /googleSheets/selectedSheet instead."
      },
      "delete": {
        "tags": [
          "Google Integrations"
        ],
        "summary": "Disconnect a Google Sheet",
        "operationId": "deleteGoogleSheetsSheet",
        "responses": {
          "200": {
            "description": "The stored sheet association was removed.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "Sheet removed"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "legacyUserJwt": []
          }
        ],
        "description": "Disconnects the Google Sheet currently associated with the App Connect user and returns plain text. It does not delete the underlying file from Google Drive."
      }
    },
    "/googleSheets/selectedSheet": {
      "post": {
        "tags": [
          "Google Integrations"
        ],
        "summary": "Select a Google Sheet tab",
        "operationId": "postGoogleSheetsSelectedSheet",
        "requestBody": {
          "$ref": "#/components/requestBodies/GenericJson"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/OkJson"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "description": "Associates an existing Google Sheet and tab selected with a Google access token. The route resolves the Google user from that token before saving the selection."
      }
    },
    "/admin/googleSheets/filePicker": {
      "get": {
        "tags": [
          "Google Integrations"
        ],
        "summary": "Open admin Google Sheets file picker",
        "operationId": "getAdminGoogleSheetsFilePicker",
        "responses": {
          "200": {
            "description": "Administrator Google Sheets picker page.",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "legacyUserJwt": [],
            "legacyRingCentralAccessToken": []
          }
        ],
        "description": "Serves the bundled administrator Google Sheets picker HTML after resolving the App Connect user. RingCentral administrator validation occurs in the subsequent configuration operations."
      }
    },
    "/admin/googleSheets/sheet": {
      "post": {
        "tags": [
          "Google Integrations"
        ],
        "summary": "Create or connect an admin Google Sheet",
        "operationId": "postAdminGoogleSheetsSheet",
        "requestBody": {
          "$ref": "#/components/requestBodies/GenericJson"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/OkJson"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "legacyUserJwt": [],
            "legacyRingCentralAccessToken": []
          }
        ],
        "description": "Creates a new Google Sheet, validates the supplied RingCentral administrator token, and stores the resulting sheet as account-level configuration."
      }
    },
    "/admin/googleSheets/selectedSheet": {
      "post": {
        "tags": [
          "Google Integrations"
        ],
        "summary": "Select an admin Google Sheet tab",
        "operationId": "postAdminGoogleSheetsSelectedSheet",
        "requestBody": {
          "$ref": "#/components/requestBodies/GenericJson"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/OkJson"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "legacyRingCentralAccessToken": []
          }
        ],
        "description": "Associates an existing Google Sheet with the RingCentral account after Google user lookup and RingCentral administrator validation."
      }
    },
    "/admin/googleSheets/config": {
      "get": {
        "tags": [
          "Google Integrations"
        ],
        "summary": "Read admin Google Sheets config",
        "operationId": "getAdminGoogleSheetsConfig",
        "responses": {
          "200": {
            "$ref": "#/components/responses/OkJson"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "legacyUserJwt": [],
            "legacyRingCentralAccessToken": []
          }
        ],
        "description": "Returns account-level Google Sheets configuration after validating the App Connect CRM user and RingCentral administrator token."
      }
    },
    "/googleDrive/oauthUrl": {
      "get": {
        "tags": [
          "Google Integrations"
        ],
        "summary": "Get Google Drive OAuth URL",
        "operationId": "getGoogleDriveOauthUrl",
        "responses": {
          "200": {
            "$ref": "#/components/responses/OkJson"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "security": [
          {
            "legacyUserJwt": []
          }
        ],
        "parameters": [
          {
            "name": "pluginId",
            "in": "query",
            "required": false,
            "description": "Plugin identifier included in OAuth state.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "description": "Returns the bundled Google Drive plugin's provider authorization URL for the App Connect user. pluginId is preserved in OAuth state for the browser callback."
      }
    },
    "/googleDrive/oauthCallback": {
      "get": {
        "tags": [
          "Google Integrations"
        ],
        "summary": "Google Drive OAuth callback",
        "operationId": "getGoogleDriveOauthCallback",
        "responses": {
          "200": {
            "$ref": "#/components/responses/OkJson"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "parameters": [
          {
            "name": "callbackUri",
            "in": "query",
            "required": true,
            "description": "Original OAuth callback URI containing encoded state.",
            "schema": {
              "type": "string",
              "format": "uri"
            }
          },
          {
            "name": "code",
            "in": "query",
            "required": true,
            "description": "Google OAuth authorization code.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "description": "Granted Google OAuth scopes.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "description": "Completes the bundled Google Drive OAuth flow by reconstructing the provider callback URI, resolving the App Connect user from state, storing Google credentials, and returning the plugin identifier."
      }
    },
    "/googleDrive/checkAuth": {
      "get": {
        "tags": [
          "Google Integrations"
        ],
        "summary": "Check Google Drive auth state",
        "operationId": "getGoogleDriveCheckAuth",
        "responses": {
          "200": {
            "$ref": "#/components/responses/OkJson"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "security": [
          {
            "legacyUserJwt": []
          }
        ],
        "description": "Checks whether the bundled Google Drive plugin has stored authorization for the App Connect user identified by the legacy session token."
      }
    },
    "/googleDrive/logout": {
      "post": {
        "tags": [
          "Google Integrations"
        ],
        "summary": "Disconnect Google Drive",
        "operationId": "postGoogleDriveLogout",
        "responses": {
          "200": {
            "$ref": "#/components/responses/OkJson"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "requestBody": {
          "$ref": "#/components/requestBodies/GoogleDriveLogout"
        },
        "description": "Disconnects the bundled Google Drive plugin user identified by the App Connect CRM-session JWT in the body. The token is sensitive and must not be logged."
      }
    },
    "/pipedrive-redirect": {
      "get": {
        "tags": [
          "Pipedrive"
        ],
        "summary": "Handle Pipedrive redirect",
        "operationId": "getPipedriveRedirect",
        "responses": {
          "200": {
            "description": "Pipedrive redirect helper page.",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "The redirect helper page could not be served."
          }
        },
        "description": "Serves the static Pipedrive redirect-helper HTML page used by the bundled connector. This route does not complete OAuth or mutate user state by itself."
      },
      "delete": {
        "tags": [
          "Pipedrive"
        ],
        "summary": "Process a Pipedrive uninstall callback",
        "operationId": "deletePipedriveRedirect",
        "responses": {
          "200": {
            "description": "Pipedrive user disconnected.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "description": "Pipedrive Basic authentication failed."
          },
          "500": {
            "description": "The Pipedrive user could not be disconnected."
          }
        },
        "security": [
          {
            "pipedriveBasicAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "user_id"
                ],
                "properties": {
                  "user_id": {
                    "type": "string",
                    "description": "Pipedrive user identifier whose CRM authorization and App Connect user record will be removed."
                  }
                },
                "additionalProperties": false,
                "description": "Pipedrive uninstall payload authenticated with the application HTTP Basic credential."
              }
            }
          }
        },
        "description": "Handles Pipedrive's authenticated uninstall callback. After HTTP Basic authentication, it invokes connector revocation for the supplied Pipedrive user_id and deletes that Pipedrive user record. Repeating the callback for a missing user still returns success."
      }
    },
    "/plugin/licenseStatus/{pluginId}": {
      "get": {
        "tags": [
          "Google Integrations"
        ],
        "summary": "Get app-specific plugin license status",
        "operationId": "getPluginLicenseStatusByPluginId",
        "parameters": [
          {
            "name": "pluginId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Identifier of a bundled, application-specific plugin supported by this server."
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/OkJson"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "legacyUserJwt": []
          }
        ],
        "description": "Returns license and authorization status for a bundled, application-specific plugin selected by pluginId. This legacy route uses userJwt in the query and is omitted from the public connector specification."
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "userJwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "App Connect CRM-session JWT issued by /apiKeyLogin or /oauth-callback. It is not a RingCentral OAuth token. A response can rotate it in X-Refreshed-JWT-Token."
      },
      "legacyUserJwt": {
        "type": "apiKey",
        "in": "query",
        "name": "jwtToken",
        "description": "Legacy query location for the App Connect CRM-session JWT. Prefer Authorization: Bearer. Query credentials can leak through URLs, logs, analytics, and browser history and must never be recorded."
      },
      "ringCentralAccessToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-RC-Access-Token",
        "description": "RingCentral OAuth access token used to validate a RingCentral user or administrator. This credential is separate from userJwt and must not be logged."
      },
      "legacyRingCentralAccessToken": {
        "type": "apiKey",
        "in": "query",
        "name": "rcAccessToken",
        "description": "Legacy query location for a RingCentral OAuth access token. Prefer X-RC-Access-Token. Query credentials can leak through URLs and logs; some legacy request bodies also carry rcAccessToken."
      },
      "mcpBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "RingCentral OAuth access token used by authenticated MCP methods."
      },
      "widgetSessionToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-App-Connect-Widget-Token",
        "description": "Short-lived signed session token issued to the App Connect MCP widget. It authenticates the allowed widget context without acting as a RingCentral OAuth token."
      },
      "filePickerJwt": {
        "type": "apiKey",
        "in": "query",
        "name": "token",
        "description": "Legacy App Connect user-session JWT used by the Google Sheets picker page."
...
  },
...
}